apps/backend/requirements.txt
splink
djangorestframework
polars
boto3
django-environ